Added logic for viewing a comment's context
Also added a simple UI element to display when you are viewing a single
comment's context.

A light-theme element needs to be added since the orange is kinda ugly
on the white background.

Also, when the 'Focus Parent' item is selected on a top-level comment,
it will focus the OP.