If the user chooses to load context for a comment...
Make sure this information is stored, so that when the comments page is
reloaded, it will reload with the proper context, rather than reloading
the whole thread.

Say you reply to a comment when in the context view, the app will now
keep the context view, rather than going back to the full thread.

Also fixed an app crash when the user clicks the 'load more comments'
button quickly. (Click one element more than once, or click two very
quickly)

Also added logic to make it so the app won't attempt to load a [deleted]
user's profile.