Added the ability to get to links from the Inbox and Profile listings.
Also, Thing contents are not parsed until the user clicks on them.
(There's no current reason to parse the comments for URLs if the user
may not ever click on them -- doing so just slowed things down,
especially in a thread with a lot of links in the comments.)

Added a common utility function in Common to display the 'Links' dialog.
(It no longer adds an onclick listener to a button argument)

Removed some unused variables from the modified files.