Adds support for the 'distinguished' field
Now comments that are 'distinguished' (moderator or admin) take
precedence over the submitter's style. This will now also show in
threads where the submitter has been distinguished. (This required
creating a separate field for the submitter's name.)

This works in the comments view, and user's profile view. The inbox is
designed differently, so it doesn't work there. Maybe in the future?

Also changed the 'when' strings from inline-constants to be localizable
strings.